pyproject.toml
[project]
name = "x-mcp"
version = "0.1.0"
description = "A server to create drafts of X(twitter) posts, threads using LLMs and post directly from the chat."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "httpx>=0.28.0",
 "mcp>=1.1.0",
 "python-dotenv>=1.0.1",
 "tweepy>=4.14.0",
]
[[project.authors]]
name = "Vidhu Panhavoor Vasudevan"
email = "vidhuvasudevan.2000@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
x-mcp = "x_mcp:main"
